Refactor all files. Reorganize the project structure with RestControllers. Add Exception Handling for REST API

+++ src/main/java/com/shantanu/blogapp/controller/CommentController.java
@@ -9,10 +9,13 @@ import com.shantanu.blogapp.service.CommentService;
 import com.shantanu.blogapp.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 
-@RestController
-@RequestMapping("/api/posts")
+@Controller
 public class CommentController {
 
 	@Autowired
@@ -24,23 +27,13 @@ public class CommentController {
 	@Autowired
 	private UserRepository userRepository;
 
-//	@PostMapping("/post/saveComment/{id}")
-//	public String saveComment(@PathVariable("id") int id, Comment comment,
-//														@AuthenticationPrincipal UserDetailsImpl currentUser) {
-//		Post post = postService.getPostById(id);
-//		post.addComment(commentService.addCommentDetails(post, comment, currentUser));
-//		postService.saveComment(post);
-//		return "redirect:/post/{id}";
-//	}
-
-	@PostMapping("/{id}/comments")
-	public void saveComment(@PathVariable("id") int postId, @RequestBody Comment comment) {
-		System.out.println(comment.getName());
-		System.out.println(comment.getText());
-		System.out.println(comment.getText());
-		Post post = postService.getPostById(postId);
-		post.addComment(commentService.addCommentDetails(post, comment));
+	@PostMapping("/post/saveComment/{id}")
+	public String saveComment(@PathVariable("id") int id, Comment comment,
+														@AuthenticationPrincipal UserDetailsImpl currentUser) {
+		Post post = postService.getPostById(id);
+		post.addComment(commentService.addCommentDetails(post, comment, currentUser));
 		postService.saveComment(post);
+		return "redirect:/post/{id}";
 	}
 
 	@GetMapping("/post/{postId}/comment/{commentId}")
@@ -60,51 +53,25 @@ public class CommentController {
 		return "updatecomment";
 	}
 
-//	@PostMapping("/post/{postId}/updateComment/{commentId}")
-//	public String updateComment(@PathVariable("commentId") int commentId,
-//															@ModelAttribute("comment") Comment comment) {
-//		Comment oldComment = commentService.getCommentById(commentId);
-//		commentService.updateComment(comment, oldComment);
-//		return "redirect:/post/{postId}";
-//	}
-
-	@PutMapping("/{postId}/comments/{commentId}")
+	@PostMapping("/post/{postId}/updateComment/{commentId}")
 	public String updateComment(@PathVariable("commentId") int commentId,
-															@RequestBody Comment comment) {
-		Comment oldComment;
-		try {
-			oldComment = commentService.getCommentById(commentId);
-		} catch (RuntimeException e) {
-			return e.getMessage();
-		}
+															@ModelAttribute("comment") Comment comment) {
+		Comment oldComment = commentService.getCommentById(commentId);
 		commentService.updateComment(comment, oldComment);
-		return "Comment updated.";
+		return "redirect:/post/{postId}";
 	}
 
-
-//	@PostMapping("/post/{postId}/deleteComment/{commentId}")
-//	public String deleteComment(@PathVariable("postId") int postId,
-//															@PathVariable("commentId") int commentId,
-//															@AuthenticationPrincipal UserDetailsImpl currentUser) {
-//		Post postById = postService.getPostById(postId);
-//		User user = userRepository.findByUsername(currentUser.getUsername()).get();
-//		if(!currentUser.getUsername().equals(postById.getAuthor()) &&  !user.getRole().equals("ROLE_ADMIN")) {
-//			throw new RuntimeException("You are not authorized to view this page");
-//		}
-//		Comment comment = commentService.getCommentById(commentId);
-//		commentService.deleteComment(comment);
-//		return "redirect:/post/{postId}";
-//	}
-
-	@DeleteMapping("/{postId}/comments/{commentId}")
-	public String deleteComment(@PathVariable("commentId") int commentId) {
-		Comment comment;
-		try {
-			comment = commentService.getCommentById(commentId);
-		} catch (RuntimeException e) {
-			return e.getMessage();
+	@PostMapping("/post/{postId}/deleteComment/{commentId}")
+	public String deleteComment(@PathVariable("postId") int postId,
+															@PathVariable("commentId") int commentId,
+															@AuthenticationPrincipal UserDetailsImpl currentUser) {
+		Post postById = postService.getPostById(postId);
+		User user = userRepository.findByUsername(currentUser.getUsername()).get();
+		if(!currentUser.getUsername().equals(postById.getAuthor()) &&  !user.getRole().equals("ROLE_ADMIN")) {
+			throw new RuntimeException("You are not authorized to view this page");
 		}
+		Comment comment = commentService.getCommentById(commentId);
 		commentService.deleteComment(comment);
-		return "Comment Deleted";
+		return "redirect:/post/{postId}";
 	}
 }
